requirements.txt
react-router-dom
flask
flask-cors
gunicorn
pandas
python-dotenv
supabase